Signal handler

+++ experimaestro/tests/test_tasks.py
@@ -6,13 +6,14 @@ from pathlib import Path
 import logging
 import pytest
 import subprocess
+import signal
 import psutil
 
 from experimaestro import *
 from experimaestro.scheduler import JobState
 from experimaestro.click import cli
 
-from .utils import TemporaryDirectory, TemporaryExperiment
+from .utils import TemporaryDirectory, TemporaryExperiment, is_posix
 
 from .tasks import *
 from . import restart
@@ -70,36 +71,55 @@ def test_done():
     pass
 
 
+TERMINATES_FUNC = [lambda p: p.terminate()]
 
-def test_restart():
+if is_posix():
+    TERMINATES_FUNC.append(lambda p: p.send_signal(signal.SIGINT))
+
+@pytest.mark.parametrize('terminate', TERMINATES_FUNC)
+def test_restart(terminate):
     """Restarting the experiment should take back running tasks"""
+    p = None
+    xpmprocess = None
+    try:
+        with TemporaryExperiment("restart", maxwait=10) as ws:
+            # Create the task and so we can get the file paths
+            task = restart.Restart()
+            task.submit(dryrun=True)
 
-    with TemporaryExperiment("restart", maxwait=10) as ws:
-        # Create the task and so we can get the file paths
-        task = restart.Restart()
-        task.submit(dryrun=True)
+            # Start the experiment with another process, and kill the job
+            command = [sys.executable, restart.__file__, ws.path]
+            logging.debug("Starting other process with: %s", command)
+            xpmprocess = subprocess.Popen(command)
+            while not task.touch.is_file():
+                time.sleep(0.1)
+            
+            pid = int(task.__xpm__.job.pidpath.read_text())
+            p = psutil.Process(pid)
+            
+            logging.debug("Process has started [file %s, pid %d]", task.touch, pid)
+            terminate(xpmprocess)
+            errorcode = xpmprocess.wait(5)
+            logging.debug("Process finishing with status %d", errorcode)
 
-        # Start the experiment with another process, and kill the job
-        command = [sys.executable, restart.__file__, ws.path]
-        logging.debug("Starting other process with: %s", command)
-        p = subprocess.Popen(command)
-        while not task.touch.is_file():
-            time.sleep(0.1)
-        logging.debug("Process has started [file %s]", task.touch)
-        p.terminate()
-        logging.debug("Process finishing with status %d", p.wait())
+            # Check that task is still running
+            logging.info("Checking that job (PID %s) is still running", pid)
+            assert p.is_running()
 
-        # Check that task is still running
-        pid = int(task.__xpm__.job.pidpath.read_text())
-        logging.info("Checking that job (PID %s) is still running", pid)
-        p = psutil.Process(pid)
-        assert p.is_running()
+            # Now, submit the job - it should pick up the process
+            # where it was left
+            logging.debug("Submitting the job")
+            Scheduler.CURRENT.submit(task.__xpm__.job)
+            with task.wait.open("w") as fp:
+                fp.write("done")
 
-        # Now, submit the job - it should pick up the process
-        # where it was left
-        logging.debug("Submitting the job")
-        Scheduler.CURRENT.submit(task.__xpm__.job)
-        with task.wait.open("w") as fp:
-            fp.write("done")
+            assert task.__xpm__.job.wait() == JobState.DONE
+    finally:
+        # Force kill
+        if xpmprocess and xpmprocess.poll() is None:
+            logging.warning("Forcing to quit process %s", xpmprocess.pid)
+            xpmprocess.kill()
 
-        assert task.__xpm__.job.wait() == JobState.DONE
+        if p and p.is_running():
+            logging.warning("Forcing to quit process %s", p.pid)
+            p.terminate()
